Rename .gitignore to avoid conflicts
Since .gitignore already means something to this repository,
we need to rename the version we want to use globally.

We do the same to .gitconfig just to be consistent. Eventually,
I'll rename all the dot files will just to increase their
visibility in the repo. 

Of course, the symbolic links to these files should keep
their leading dot in their names.